requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==3.4
Django==4.0.4
django-encrypted-model-fields==0.6.5
idna==3.3
psycopg2==2.9.3
pycparser==2.21
requests==2.27.1
semantic-version==2.10.0
setuptools-rust==1.3.0
sqlparse==0.4.2
typing-extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
